part a solved

+++ src/main/kotlin/AdventDay.kt
@@ -3,8 +3,8 @@ package org.example
 import java.io.File
 
 abstract class AdventDay(open val inputFilePath: String) {
-    abstract fun solveA(): Int
-    abstract fun solveB(): Int
+    abstract fun solveA(): Long
+    abstract fun solveB(): Long
 
     protected fun getAllFilesLines() = File(inputFilePath).useLines { it.toList() }
